src/hooks/usePagination/usePagination.ts
import { useState, useMemo } from "react";
import { APP } from "@/constants";

type TUsePaginationProps<T> = {
  items: T[];
  itemsPerPage?: number;
};

export function usePagination<T>({
  items,
  itemsPerPage = APP.POKEMON.PAGINATION.ITEMS_PER_PAGE,
}: TUsePaginationProps<T>) {
  const [currentPage, setCurrentPage] = useState(1);

  const totalPages = Math.ceil(items.length / itemsPerPage);

  const paginatedItems = useMemo(() => {
    const start = (currentPage - 1) * itemsPerPage;
    const end = start + itemsPerPage;
    return items.slice(start, end);
  }, [items, currentPage, itemsPerPage]);

  const goToNextPage = () => {
    if (currentPage < totalPages) {
      setCurrentPage((prev) => prev + 1);
    }
  };

  const goToPreviousPage = () => {
    if (currentPage > 1) {
      setCurrentPage((prev) => prev - 1);
    }
  };

  return {
    currentPage,
    totalPages,
    paginatedItems,
    goToNextPage,
    goToPreviousPage,
  };
}
